Fixing things up
