+++ setup.py
@@ -114,9 +114,4 @@ setuptools.setup(
         "Operating System :: OS Independent",
     ],
     python_requires=">=3.6",
-        entry_points={
-        'console_scripts': [
-            'bolt-get-hooks=cli.get_hooks:main',
-        ]
-    },
 )
