+++ slack_bolt/cli/get_hooks.py
@@ -8,7 +8,7 @@ hook_payload = {
         "start": "python -m slack_bolt.cli.start",
     },
     "config": {
-        "watch": {"filter-regex": "^manifest\\.(json)$", "paths": ["."]},
+        "watcher": {"filter-regex": "^manifest\\.(json)$", "paths": ["."]},
         "sdk-managed-connection-enabled": True,
     },
 }
