+++ slack_bolt/cli/get_manifest.py
@@ -2,12 +2,9 @@
 import os
 import re
 from .error import CliError
-from .utils import handle_exceptions
 from typing import List
 
-FILE = "manifest"
-
-EXPLICIT_EXCLUDED_DIRECTORIES = [
+EXCLUDED_DIRECTORIES = [
     "lib",
     "bin",
     "include",
@@ -22,7 +19,7 @@ EXPLICIT_EXCLUDED_DIRECTORIES = [
     "tests",
 ]
 
-DIRECTORY_IGNORE_REGEX = re.compile(r"(^\.|^\_|^{}$)".format("$|^".join(EXPLICIT_EXCLUDED_DIRECTORIES)), re.IGNORECASE)
+DIRECTORY_IGNORE_REGEX = re.compile(r"(^\.|^\_|^{}$)".format("$|^".join(EXCLUDED_DIRECTORIES)), re.IGNORECASE)
 
 
 def filter_directories(directories: List[str]) -> List[str]:
@@ -37,15 +34,18 @@ def find_file_path(path: str, file: str) -> str:
     raise CliError(f"Manifest file not found!\nPath: {path}\nFile: {file}")
 
 
-@handle_exceptions()
 def get_manifest(working_directory: str) -> str:
-    file_path = find_file_path(working_directory, f"{FILE}.json")
+    file_path = find_file_path(working_directory, f"manifest.json")
 
-    print(file_path)
     with open(file_path, "r") as manifest:
         return manifest.read()
 
 
 if __name__ == "__main__":
-    current_wd = os.getcwd()
-    print(get_manifest(current_wd))
+    try:
+        print(get_manifest(os.getcwd()))
+    except CliError as e:
+        print(e)
+        exit()
+    except KeyboardInterrupt:
+        exit(130)
