+++ slack_bolt/cli/start.py
@@ -1,9 +1,11 @@
 #!/usr/bin/env python
 import os
 import runpy
+import sys
+
+from slack_bolt.app.app import App
 
 from .error import CliError
-from .utils import load_app_module, handle_exceptions
 
 
 DEFAULT_ENTRYPOINT_FILE = "app.py"
@@ -22,13 +24,56 @@ def validate_env() -> None:
 
 
 def get_entrypoint_path(working_directory: str) -> str:
-    custom_path = os.environ.get("SLACK_CLI_CUSTOM_FILE_PATH", None)
+    custom_path = os.environ.get("SLACK_APP_PATH", None)
     if custom_path:
         return f"{working_directory}/{custom_path}"
     return f"{working_directory}/{DEFAULT_ENTRYPOINT_FILE}"
 
 
-@handle_exceptions()
+def load_app_module(entrypoint_path: str) -> str:
+    module_name = _get_module_name(entrypoint_path)
+
+    try:
+        __import__(module_name)
+    except ImportError:
+        raise CliError(f"Unable to load module {module_name} found in Entrypoint {entrypoint_path}!")
+
+    module = sys.modules[module_name]
+
+    matches = [v for v in module.__dict__.values() if isinstance(v, App)]
+
+    if len(matches) >= 1:
+        return module_name
+
+    raise CliError(f"No Slack App(s) objects detected in module {module_name} from entrypoint {entrypoint_path}")
+
+
+def _get_module_name(path: str) -> str:
+    path = os.path.realpath(path)
+
+    fname, ext = os.path.splitext(path)
+    if ext == ".py":
+        path = fname
+
+    if os.path.basename(path) == "__init__":
+        path = os.path.dirname(path)
+
+    module_name = []
+
+    # move up until outside package structure (no __init__.py)
+    while True:
+        path, name = os.path.split(path)
+        module_name.append(name)
+
+        if not os.path.exists(os.path.join(path, "__init__.py")):
+            break
+
+    if sys.path[0] != path:
+        sys.path.insert(0, path)
+
+    return ".".join(module_name[::-1])
+
+
 def start(working_directory: str) -> None:
     validate_env()
 
@@ -50,5 +95,10 @@ def start(working_directory: str) -> None:
 
 
 if __name__ == "__main__":
-    current_wd = os.getcwd()
-    start(current_wd)
+    try:
+        start(os.getcwd())
+    except CliError as e:
+        print(e)
+        exit()
+    except KeyboardInterrupt:
+        exit(130)
